separate data loader and mnist loader
