+++ src/index.js
@@ -3,3 +3,4 @@
 exports.layers = require('./layers');
 exports.Network = require('./network');
 exports.MnistLoader = require('./mnist_loader');
+exports.DataLoader = require('./data_loader');
